+++ .github/workflows/unity-ci.yml
@@ -1,66 +1,50 @@
+
 name: Unity CI
 
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-  release:
-    types: [published]
+on: [push, pull_request]
 
 jobs:
-  build:
+  test:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
+    - name: Checkout
+      uses: actions/checkout@v2
 
-      - name: Set up Unity
-        uses: game-ci/unity-setup@latest
-        with:
-          unityVersion: "2021.3.41f1"
+    - name: Cache
+      uses: actions/cache@v2
+      with:
+        path: Library
+        key: ${{ runner.os }}-library-${{ hashFiles('**/ProjectSettings/ProjectVersion.txt') }}
+        restore-keys: |
+          ${{ runner.os }}-library-
 
-      - name: Build project
-        uses: game-ci/unity-builder@v2
-        with:
-          projectPath: .
-          unityVersion: "2021.3.41f1"
+    - name: Setup Unity
+      uses: game-ci/unity-action@v2
+      with:
+        unityVersion: 6000.0.24f1
 
-      - name: Run tests
-        uses: game-ci/unity-test-runner@v2
-        with:
-          projectPath: .
-          unityVersion: "2021.3.41f1"
+    - name: Run Tests
+      uses: game-ci/unity-test-runner@v2
+      with:
+        unityVersion: 6000.0.24f1
+        testMode: playmode
+        artifactsPath: artifacts
 
-  publish:
-    needs: build
-    runs-on: ubuntu-latest
-    if: github.event_name == 'release'
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
+    - name: Upload Artifacts
+      if: always()
+      uses: actions/upload-artifact@v2
+      with:
+        name: Test Artifacts
+        path: artifacts
 
-      - name: Set up Unity
-        uses: game-ci/unity-setup@v2
-        with:
-          unityVersion: "2021.3.41f1"
+    - name: Build
+      uses: game-ci/unity-builder@v2
+      with:
+        unityVersion: 6000.0.24f1
+        targetPlatform: StandaloneWindows64
 
-      - name: Build project
-        uses: game-ci/unity-builder@v2
-        with:
-          projectPath: .
-          unityVersion: "2021.3.41f1"
-
-      - name: Create Unity package
-        run: |
-          mkdir -p output
-          cp -r Assets/YourPackageName output/
-
-      - name: Publish Unity package to GitHub Releases
-        uses: softprops/action-gh-release@v1
-        with:
-          files: output/*
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+    - name: Upload Build
+      uses: actions/upload-artifact@v2
+      with:
+        name: Build
+        path: build
